src/remote-warehouse/requirements.txt
confluent-kafka==2.3.0
protobuf==4.25.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation-logging==0.43b0
grpcio-tools==1.60.0
